feat: useVisibility hook
