+++ src/hooks/useVisibility.ts
@@ -1,0 +1,30 @@
+import { useEffect, useState } from 'react';
+
+const useVisibility = (
+  onClose: () => void,
+  onConfirm?: () => void,
+  initialState = false,
+) => {
+  const [isVisible, setIsVisible] = useState(initialState);
+
+  useEffect(() => {
+    setIsVisible(true);
+  }, []);
+
+  const handleClose = () => {
+    setIsVisible(false);
+    setTimeout(onClose, 300);
+  };
+
+  if (onConfirm) {
+    const handleConfirm = () => {
+      setIsVisible(false);
+      setTimeout(onConfirm, 300);
+    };
+    return { isVisible, handleClose, handleConfirm };
+  }
+
+  return { isVisible, handleClose };
+};
+
+export default useVisibility;
